+++ info.rkt
@@ -2,7 +2,7 @@
 
 (define collection "widgetkit")
 (define scribblings '(("widgetkit.scrbl" ())))
-(define version "0.7.1")
+(define version "0.7.2")
 (define pkg-desc "A curated collection of GUI widgets for Racket")
 (define pkg-authors '("turinglambdaai"))
 (define license 'MIT)